+++ pages/formula.php
@@ -51,9 +51,9 @@ $base_ex = get_formula_excludes($conn, $fid, 'base');
 <style>
 .mfp-iframe-holder .mfp-content {
     line-height: 0;
-    width: 1100px;
-    max-width: 1100px; 
-	height: 1100px;
+    width: 1450px;
+    max-width: 1450px; 
+	height: 1300px;
 }
 </style>
 <div id="content-wrapper" class="d-flex flex-column">