+++ pages/listIngredients.php
@@ -189,7 +189,7 @@ $next = $page + 1;
         </nav>
 <script type="text/javascript" language="javascript" >
 $(".alert-dismissible").fadeTo(2000, 500).slideUp(500, function(){
-    $(".alert-dismissible").slideUp(500);
+    $(".alert-dismissible").alert('close');
 });
 
 $('a[rel=tipsy]').tipsy();