+++ pages/listIngredientsCards.php
@@ -160,8 +160,9 @@ $next = $page + 1;
    </div>
 
 <script type="text/javascript" language="javascript" >
+
 $(".alert-dismissible").fadeTo(2000, 500).slideUp(500, function(){
-    $(".alert-dismissible").slideUp(500);
+    $(".alert-dismissible").alert('close');
 });
 
 $('a[rel=tipsy]').tipsy();